+++ EssentialFeed/EssentialFeed/Feed Cache/FeedStore.swift
@@ -1,0 +1,13 @@
+//
+//  Copyright © 2019 Essential Developer. All rights reserved.
+//
+
+import Foundation
+
+public protocol FeedStore {
+	typealias DeletionCompletion = (Error?) -> Void
+	typealias InsertionCompletion = (Error?) -> Void
+	
+	func deleteCachedFeed(completion: @escaping DeletionCompletion)
+	func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
